+++ gestao_residuos/src/main/java/br/com/fiap/gestao_residuos/dto/LocalColetaExibicaoDTO.java
@@ -2,13 +2,14 @@ package br.com.fiap.gestao_residuos.dto;
 
 import br.com.fiap.gestao_residuos.model.LocalColeta;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record LocalColetaExibicaoDTO(
         Long id,
         String nome,
         String endereco,
-        Double capacidadeMaxima
+        BigDecimal capacidadeMaxima
 ) {
     public LocalColetaExibicaoDTO(LocalColeta localColeta){
         this(
