+++ gestao_residuos/src/main/java/br/com/fiap/gestao_residuos/dto/ResiduoCadastroDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record ResiduoCadastroDTO(
 
         Long id,
@@ -16,6 +18,6 @@ public record ResiduoCadastroDTO(
 
         @NotNull(message = "Peso é obrigatório")
         @Positive(message = "Peso deve ser maior que zero")
-        Double peso
+        BigDecimal peso
 ) {
 }
